fix(#44): move deployment_profile.py from core/ to contracts/
Pure StrEnum + brick constants + frozen dataclass, no kernel deps.
- Move core/deployment_profile.py → contracts/deployment_profile.py
- Update 15 caller files (23 import sites total)
