+++ tests/unit/core/test_deployment_profile.py
@@ -16,7 +16,7 @@ import logging
 
 import pytest
 
-from nexus.core.deployment_profile import (
+from nexus.contracts.deployment_profile import (
     ALL_BRICK_NAMES,
     BRICK_A2A,
     BRICK_CACHE,
