+++ tests/unit/core/test_factory_boot.py
@@ -16,6 +16,7 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
+from nexus.contracts.deployment_profile import DeploymentProfile
 from nexus.contracts.exceptions import BootError
 from nexus.core.config import (
     BrickServices,
@@ -24,7 +25,6 @@ from nexus.core.config import (
     PermissionConfig,
     SystemServices,
 )
-from nexus.core.deployment_profile import DeploymentProfile
 from nexus.core.performance_tuning import resolve_profile_tuning
 from nexus.factory import (
     _boot_brick_services,
