+++ tests/unit/test_factory.py
@@ -17,8 +17,8 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
+from nexus.contracts.deployment_profile import DeploymentProfile
 from nexus.contracts.exceptions import BootError, NexusError
-from nexus.core.deployment_profile import DeploymentProfile
 
 # ---------------------------------------------------------------------------
 # TestBootError
